requirements.txt
torch==2.2.2
torchvision
torchaudio
sentence-transformers==2.2.2
huggingface-hub==0.16.4
transformers==4.30.2
numpy<2
gradio==4.19.2
pillow
